Revert "Poller: add 24 bit PID size support, fix routine control PID sizes"
This reverts commit caf65c23ce6a4961d8e4197de4c1334c81582860.

+++ vehicle/OVMS.V3/components/vehicle_cadillac_c2_cts/src/vehicle_cadillac_c2_cts.cpp
@@ -240,7 +240,7 @@ ESP_LOGI(TAG, "now awake");
 
 void
 OvmsVehicleCadillaccC2CTS::IncomingPollReply(canbus* bus, uint16_t type,
-  uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
+  uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain)
   {
   int value1 = (int)data[0];
   int value2 = ((int)data[0] << 8) + (int)data[1];
