+++ vehicle/OVMS.V3/components/vehicle_vweup/src/vweup_t26.h
@@ -183,7 +183,7 @@ class OvmsVehicleVWeUpAll : public OvmsVehicle
 
     //OBD
     void ObdInit();
-    void IncomingPollReply(canbus *bus, uint16_t type, uint32_t pid, uint8_t *data, uint8_t length, uint16_t mlremain);
+    void IncomingPollReply(canbus *bus, uint16_t type, uint16_t pid, uint8_t *data, uint8_t length, uint16_t mlremain);
 
 //  protected:
 //    virtual void Ticker1(uint32_t ticker);
